requirements.txt
# Core Framework
fastapi==0.121.0
uvicorn[standard]==0.38.0
streamlit==1.51.0

# AI & LLM
google-genai==1.49.0
langchain==1.0.5
langchain-google-genai==3.0.1
langgraph==1.0.2

# Vector Database & Embeddings
qdrant-client==1.15.1
sentence-transformers==5.1.2
torch==2.9.0
transformers==4.57.1

# Database
pymongo==4.15.3
motor==3.7.1

# Document Processing
pypdf==6.1.3
Pillow==12.0.0
python-multipart==0.0.20

# Research Tools
arxiv==2.3.0
httpx==0.28.1

# Utilities
# python-dotenv==0.0.1
pydantic==2.12.4
pydantic-settings==2.11.0

# Logging & Monitoring
loguru==0.7.3
